feat: add launch setup
